Ubifs image detection
Version 1 of ubifs detection. After finding initial offset
(maybe refactored to regex find) skip to last node-head and fetch length
from node. Assuming this is the final node there is an additional fill
with an offset of 4 bytes after the nodesize field.
The above method ignores the possibility of another fs inside the
expected ubifs image.